+++ HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## UNRELEASED
+
+* Flask rules: add support for flat domain package structure.
+
 ## 0.2.0 (2021-05-07)
 
 * Replace discover_flask() with get_flask_rules()